requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-CORS==4.0.0
pyjwt==2.8.0
cryptography==41.0.7
schedule==1.2.0
pillow==10.0.1